Move ValidationPipe to its own module & expose as service
