+++ src/core/validation/index.ts
@@ -1,0 +1,2 @@
+export * from './validation.exception';
+export * from './validation.pipe';
